Extract on_page_layout callbacks into module
In order to also run callbacks in the page edit preview we extract it into its own module.
