+++ lib/alchemy/engine.rb
@@ -40,6 +40,7 @@ require_relative './modules'
 require_relative './mount_point'
 require_relative './name_conversions'
 require_relative './on_page_layout'
+require_relative './on_page_layout/callbacks_runner'
 require_relative './page_layout'
 require_relative './permissions'
 require_relative './picture_attributes'
